Delete Income updated to check for validation and define income

+++ backend/controllers/income.js
@@ -41,6 +41,7 @@ exports.addIncome = async (req, res) => {
     console.log(income)
 }
 
+//Get Incomes
 exports.getIncomes = async (req, res) => {
     try { 
         const incomes= await incomeSchema.find().sort({createdAt: -1})
@@ -50,13 +51,26 @@ exports.getIncomes = async (req, res) => {
     }
 }
 
+//Delete Income
+exports.validateDeleteIncome =[
+    param('id').isMongoId().withMessage('Invalid ID format')
+]
+
 exports.deleteIncome = async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({errors: errors.array() });
+    }
+
     const {id} = req.params;
-    incomeSchema.findByIdAndDelete(id)
-        .then((income) => {
-            res.status(200).json({message: 'Income Deleted'})
-        })
-        .catch ((err) =>{
-            res.status(500).json({message: 'Internal Server Error', error: error.message})
-        })
-}
+
+    try {
+        const income = await incomeSchema.findByIdAndDelete(id);
+        if (!income) {
+            return res.status(404).json({message: 'Income not found! '});
+        }
+        res.status(200).json({message: 'Income Deleted'});
+    } catch (error) {
+         res.status(500).json({message: 'Internal Server Error', error: error.message});
+        }
+};
